feat: handle NPE and allow user pass no project name

+++ krell.sh
@@ -4,7 +4,7 @@ rm -rf krell-template-runner
 git clone https://github.com/ampersanda/krell-template-runner
 cd krell-template-runner
 
-if ! type bb > /dev/null; then
+if ! type bb >/dev/null; then
   echo "👉 Babashka is not installed. Installing babashka..."
   bash <(curl -s https://raw.githubusercontent.com/borkdude/babashka/master/install)
   echo "👉 Babashka installed."
@@ -14,7 +14,11 @@ all_args=("$@")
 first_arg=$1
 rest_args=("${all_args[@]:1}")
 
-bb runner.clj $1 "${rest_args[@]}"
-mv $1 ..
-cd ..
-rm -rf krell-template-runner
+bb runner.clj "${first_arg}" "${rest_args[@]}"
+
+if [[ $# -eq 0 ]]; then
+  mv ${first_arg} ..
+  cd ..
+  rm -rf krell-template-runner
+  exit 0
+fi
